### 2022-07-18 bcd3967
add joss paper generator action
