### 2022-07-25 92fe981
Added affiliation index to paper for JOSS
This should fix the error over at https://github.com/openjournals/joss-reviews/issues/4584#issuecomment-1186925516
