### 2022-07-25 #4584 @Kevin-Mattheus-Moerman
@SuperKogito This pull request should hopefully fix it: https://github.com/SuperKogito/spafe/pull/41

I just added `affiliation: 1` [as per our example](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).
