### 2022-07-25 #4584 @Kevin-Mattheus-Moerman
@SuperKogito in addition can you make sure to have a _Statement of need_ section? This is a requirement now for JOSS. Can you [check our guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) and rework your paper to cover these requirements? Thanks.
