### 2022-07-26 1b8d71a
rework paper to fit joss's requirements
