### 2022-07-26 2a024bf
fix typo as mentioned in joss's review thread
