### 2022-07-26 #4584 @SuperKogito
Just restructured the paper to comply with the requirements. The paper is yet a bit over 1k characters, due to some theory details that I deem useful. However, if the reviewers/ editor disagree, we can discuss how to  shorten it.
