### 2022-08-16 #4584 @faroit
👋 Hi @SuperKogito, thanks for your submission to JOSS, and sorry for the delay (I was on vacation last week)! I will be serving as the editor for this. We'll use this issue to identify reviewers and resolve technical issues.

@SuperKogito you already mentioned a couple of potential reviewers: pzelasko, hadware, ebezzam. I will ask them shortly.

Furthermore, I would also suggest the following potential reviewers:

- nils-werner
- hendriks73
- csteinmetz1
- ws-choi
- jonashaag

Please check these potential reviewers with respect to the JOSS [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy).
