### 2022-08-22 #4584 @faroit
👋 @nils-werner @hendriks73 @csteinmetz1 @ws-choi @jonashaag @hbredin - would any of you be willing to review this submission for JOSS? The JOSS review process takes place on GitHub and focuses on the software and a short paper. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under review is [spafe](https://github.com/SuperKogito/spafe).
