### 2022-08-23 #4584 @hbredin
I'd like to say yes because I really like the very idea of JOSS.
However, I am not sure I can commit to working on this before the end of September.
Let me know if that is OK.
