### 2022-09-04 #4739 @editorialbot
[REVIEW]: Spafe: Simplified python audio features extraction
**Submitting author:** @SuperKogito (<a href="http://orcid.org/0000-0002-9008-7562">Ayoub Malek</a>)
**Repository:** https://github.com/SuperKogito/spafe
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.1
