### 2022-09-28 #4739 @hadware
I started reading the code and the paper, but I'm currently on holidays. I'll start giving some feedback in the first week of october as well.
