### 2022-09-29 5ae1983
Streamlined dependencies and setup.py, straightened out the github actions a bit.
