+++ docs/requirements.txt
@@ -1,3 +1,0 @@
-sphinxcontrib-napoleon==0.7
-nbsphinx==0.8.9
-pydata-sphinx-theme==0.8.1
