+++ .github/workflows/ci.yml
@@ -36,14 +36,18 @@ jobs:
     steps:
       - uses: actions/checkout@v2
 
+      - name: Setup python
+        uses: actions/setup-python@v2
+        with:
+          python-version: "3.10"
+
       - name: Setup black environment
-        run: conda create --quiet --name black pyflakes
+        run: |
+          pip install --upgrade pip
+          pip install black
 
       - name: Check formatting with black
         run: |
-          export PATH="/usr/share/miniconda/bin:$PATH"
-          source activate black
-          pip install black
           black --check spafe
 
   testing:
@@ -55,14 +59,18 @@ jobs:
 
     steps:
       - uses: actions/checkout@v2
+
+      - name: Setup python
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+
       - name: Setup testing environment
-        run: conda create --quiet --name testing pytest
+        run: |
+          pip install --upgrade pip
+          pip install .[tests]
 
       - name: Test
         run: |
-          export PATH="/usr/share/miniconda/bin:$PATH"
-          source activate testing
-          pip install .
-          pip install -r tests/test-requirements.txt
           pytest -x --cache-clear -n 8 -q --cov=./ || --force-yes
           codecov
