### 2022-09-29 9d11b30
Relaxing dependencies for numpy/scipy and removing tests on 3.5-6
