+++ requirements.txt
@@ -1,3 +1,3 @@
-numpy==1.22.0
-scipy==1.8.0
-matplotlib==3.5.2
+numpy>=1.21
+scipy>=1.8.0
+matplotlib>=3.5.0
