### 2022-09-29 3013763
Setting matplotlib as an optional dependency, fixing doc conf.py .
