+++ setup.py
@@ -52,13 +52,17 @@ setup(
             "coveralls",
             "pytest-xdist",
             "codacy-coverage",
-            "matplotlib==3.5.2",
+            "matplotlib",
             "mock==4.0.3"
         ],
         "docs": [
             "sphinxcontrib-napoleon==0.7",
             "nbsphinx==0.8.9",
-            "pydata-sphinx-theme==0.8.1"
-        ]
+            "pydata-sphinx-theme==0.8.1",
+            "matplotlib"
+        ],
+        "plotting": {
+            "maplotlib"
+        }
     }
 )
