+++ spafe/frequencies/dominant_frequencies.py
@@ -9,7 +9,6 @@
 """
 import scipy
 import numpy as np
-import matplotlib.pyplot as plt
 from ..utils.preprocessing import framing, windowing
 
 
