+++ requirements.txt
@@ -1,2 +1,2 @@
 numpy>=1.21
-scipy>=1.8.0
+scipy>=1.7.3
