### 2022-09-30 9d0e0e4
Started type hinting
