+++ spafe/utils/filters.py
@@ -6,11 +6,15 @@
   For a copy, see <https://github.com/SuperKogito/spafe/blob/master/LICENSE>.
 
 """
+from typing_extensions import Literal
+
 import numpy as np
 from scipy import signal
 
+ScaleType = Literal["ascendant", "descendant", "constant"]
 
-def rasta_filter(x):
+
+def rasta_filter(x: np.ndarray) -> np.ndarray:
     """
     Implementing the RASTA filter as in [Ellis-plp]_.
 
@@ -26,7 +30,7 @@ def rasta_filter(x):
         - cols of x = critical bands
     """
     numer = np.arange(-2, 3)
-    numer = (-1 * numer) / np.sum(numer**2)
+    numer = (-1 * numer) / np.sum(numer ** 2)
     denom = np.array([1, -0.94])
 
     z = signal.lfilter_zi(numer, 1)
@@ -38,12 +42,12 @@ def rasta_filter(x):
         y1 = y1 * 0
 
         # IIR
-        y2, _ = signal.lfilter(numer, denom, x[i, 4 : x.shape[1]], axis=0, zi=z)
+        y2, _ = signal.lfilter(numer, denom, x[i, 4: x.shape[1]], axis=0, zi=z)
         y[i, :] = np.append(y1, y2)
     return y
 
 
-def scale_fbank(scale, nfilts):
+def scale_fbank(scale: ScaleType, nfilts: int) -> np.ndarray:
     """
     Generate scaling vector.
 
