### 2022-09-30 5251954
More type hinting and comments
