+++ spafe/features/spfeats.py
@@ -7,12 +7,33 @@
   For a copy, see <https://github.com/SuperKogito/spafe/blob/master/LICENSE>.
 
 """
-import scipy
+from typing import Dict
+
 import numpy as np
 from scipy import stats
 
+from typing_extensions import TypedDict
 
-def spectral_centroid(sig, fs, spectrum, i=1):
+
+class SpectralFeats(TypedDict):
+    spectral_skewness: float
+    spectral_kurtosis: float
+    spectral_entropy: float
+    spectral_spread: float
+    spectral_flatness: float
+    spectral_rolloff: float
+    spectral_flux: float
+    spectral_mean: float
+    spectral_rms: float
+    spectral_std: float
+    spectral_variance: float
+
+
+def spectral_centroid(sig,
+                      fs: int,
+                      spectrum: np.ndarray,
+                      i: int = 1) -> float:
+    # TODO: unused sig argument?
     """
     Compute the spectral centroid (which is the barycenter of the spectrum) as
     described in [Peeters]_.
@@ -31,7 +52,7 @@ def spectral_centroid(sig, fs, spectrum, i=1):
         .. math::
 
             \mu_{i} &= \\frac{\sum_{n=0}^{N}f_{k}^{i}*a_{k}}{\sum_{n=0}^{N}a_{k}}\\\\
-            S_{centroid} &= \mu_{1}\\\\
+            S_{centroid} &= \mu_{i}\\\\
     """
     # compute magnitude spectrum
     magnitude_spectrum = np.abs(spectrum)
@@ -40,11 +61,11 @@ def spectral_centroid(sig, fs, spectrum, i=1):
     freqs = np.abs(np.fft.fftfreq(len(spectrum), 1.0 / fs))
 
     # return weighted mean
-    sc = np.sum(magnitude_spectrum * freqs**i) / np.sum(magnitude_spectrum)
+    sc = np.sum(magnitude_spectrum * freqs ** i) / np.sum(magnitude_spectrum)
     return sc
 
 
-def spectral_skewness(sig, fs, spectrum):
+def spectral_skewness(sig: np.ndarray, fs: int, spectrum: np.ndarray):
     """
     Compute the spectral skewness (which is a measure of the asymmetry of a
     distribution around its mean) as described in [Peeters]_.
@@ -72,12 +93,12 @@ def spectral_skewness(sig, fs, spectrum):
 
     # return weighted mean
     sk = np.sum(magnitude_spectrum * (freqs - mu1) ** 3) / (
-        np.sum(magnitude_spectrum) * mu2**3
+            np.sum(magnitude_spectrum) * mu2 ** 3
     )
     return sk
 
 
-def spectral_kurtosis(sig, fs, spectrum):
+def spectral_kurtosis(sig: np.ndarray, fs: int, spectrum: np.ndarray) -> float:
     """
     Compute the spectral kurtosis (which is a measure of the flatness of a
     distribution around its mean) as described in [Peeters]_.
@@ -105,12 +126,13 @@ def spectral_kurtosis(sig, fs, spectrum):
 
     # return weighted mean
     sk = np.sum(magnitude_spectrum * (freqs - mu1) ** 4) / (
-        np.sum(magnitude_spectrum) * mu2**4
+            np.sum(magnitude_spectrum) * mu2 ** 4
     )
     return sk
 
 
-def spectral_entropy(sig, fs, spectrum):
+def spectral_entropy(sig: np.ndarray, fs: int, spectrum: np.ndarray) -> float:
+    # TODO: unused sig and fs arguments
     """
     Compute the spectral entropy as described in [Misra]_.
 
@@ -120,7 +142,7 @@ def spectral_entropy(sig, fs, spectrum):
[…]
