### 2022-09-30 a6fa620
More type hints and comments in features
