+++ spafe/features/pncc.py
@@ -6,15 +6,19 @@
   For a copy, see <https://github.com/SuperKogito/spafe/blob/master/LICENSE>.
 
 """
+from typing import Optional
+
 import numpy as np
 from scipy.fftpack import dct
-from ..utils.cepstral import normalize_ceps, lifter_ceps
+from ..utils.cepstral import normalize_ceps, lifter_ceps, NormalizationType
+from ..utils.converters import ErbConversionApproach
 from ..utils.exceptions import ParameterError, ErrorMsgs
 from ..fbanks.gammatone_fbanks import gammatone_filter_banks
-from ..utils.preprocessing import pre_emphasis, framing, windowing, zero_handling
+from ..utils.filters import ScaleType
+from ..utils.preprocessing import pre_emphasis, framing, windowing, zero_handling, WindowType
 
 
-def medium_time_power_calculation(P, M=2):
+def medium_time_power_calculation(P: np.ndarray, M: int = 2) -> np.ndarray:
     """
     Compute the medium time power calulations according to [Kim]_ .
 
@@ -36,7 +40,7 @@ def medium_time_power_calculation(P, M=2):
         [
             [
                 (1 / (2 * M + 1))
-                * sum(P[max(0, m - M) : min(P.shape[0], m + M + 1), l])
+                * sum(P[max(0, m - M): min(P.shape[0], m + M + 1), l])
                 for l in range(P.shape[1])
             ]
             for m in range(P.shape[0])
@@ -45,7 +49,9 @@ def medium_time_power_calculation(P, M=2):
     return Q_tilde
 
 
-def asymmetric_lowpass_filtering(Q_tilde_in, lm_a=0.999, lm_b=0.5):
+def asymmetric_lowpass_filtering(
+        Q_tilde_in: np.ndarray,
+        lm_a: float = 0.999, lm_b: float = 0.5) -> np.ndarray:
     """
     Apply asymmetric lowpass filter according to [Kim]_ .
 
@@ -69,10 +75,10 @@ def asymmetric_lowpass_filtering(Q_tilde_in, lm_a=0.999, lm_b=0.5):
     """
     Q_tilde_out = np.zeros_like(Q_tilde_in)
     Q_tilde_out[0,] = (
-        0.9
-        * Q_tilde_in[
-            0,
-        ]
+            0.9
+            * Q_tilde_in[
+                0,
+            ]
     )
 
     # compute asymmetric nonlinear filter
@@ -91,7 +97,8 @@ def asymmetric_lowpass_filtering(Q_tilde_in, lm_a=0.999, lm_b=0.5):
     return Q_tilde_out
 
 
-def temporal_masking(Q_tilde_0, lam_t=0.85, myu_t=0.2):
+def temporal_masking(Q_tilde_0: np.ndarray,
+                     lam_t: float = 0.85, myu_t: float = 0.2) -> np.ndarray:
     """
 
     Args:
@@ -131,7 +138,11 @@ def temporal_masking(Q_tilde_0, lam_t=0.85, myu_t=0.2):
     return Q_tilde_tm
 
 
-def weight_smoothing(R_tilde, Q_tilde, nfilts=128, N=4):
+def weight_smoothing(R_tilde: np.ndarray,
+                     Q_tilde: np.ndarray,
+                     nfilts: int = 128,
+                     N: int = 4) -> np.ndarray:
+    # TODO: missing argument in docstring : N
     """
     Apply spectral weight smoothing according to [Kim]_.
 
@@ -165,7 +176,10 @@ def weight_smoothing(R_tilde, Q_tilde, nfilts=128, N=4):
     return S_tilde
 
 
-def mean_power_normalization(T, lam_myu=0.999, nfilts=80, k=1):
+def mean_power_normalization(T: np.ndarray,
[…]
