### 2022-09-30 3928a60
More type hinting, should be mostly done now.
