### 2022-09-30 1760292
And again, more type hinting!
