+++ v0.1.2/_sources/examples/features_ngcc_example.ipynb.txt
@@ -1,93 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "code",
-   "execution_count": 2,
-   "metadata": {},
-   "outputs": [
-    {
-     "data": {
-      "image/png": "iVBORw0KGgoAAAANSUhEUgAAAY4AAAEKCAYAAAAFJbKyAAAABHNCSVQICAgIfAhkiAAAAAlwSFlzAAALEgAACxIB0t1+/AAAADl0RVh0U29mdHdhcmUAbWF0cGxvdGxpYiB2ZXJzaW9uIDMuMC4zLCBodHRwOi8vbWF0cGxvdGxpYi5vcmcvnQurowAAIABJREFUeJzsvWmMbdl137f2nceaX9WrN7/X/Xoi2c3JpBDJAh06kmjAZpxIgoJ8oBgC9AfJMewEEBUIESEJAgMENogAEcJEDMjEFqVosBq2LJlWTH2ITIpqUpSb3eyJ/eZ6NVfded75ULfe+q399q6uDvIoulAbaPR5t86955x99l7jf/2X897L6Tgdp+N0nI7TcdyR+au+gdNxOk7H6Tgd/2GNU8VxOk7H6Tgdp+NtjVPFcTpOx+k4HafjbY1TxXE6TsfpOB2n422NU8VxOk7H6Tgdp+NtjVPFcTpOx+k4HafjbY1Hqjicc//QOfdt59yLzrnfcM6VnHNXnXNfc8695pz7TedcYXpucfrv16d/v4Lf+fnp56845370Ud7z6Tgdp+N0nI6jxyNTHM658yLyX4vI+7337xSRrIj8lIj8DyLyT7z310VkV0Q+Mf3KJ0Rk13v/uIj8k+l54px7Zvq9d4jIj4nI/+ycyz6q+z4dp+N0nI7TcfR41KGqnIiUnXM5EamIyJqI/Mci8tvTv39BRP7T6fFHp/+W6d8/7Jxz08+/5L3ve+/fFJHXReQDj/i+T8fpOB2n43QkRu5R/bD3/q5z7n8UkVsi0hWRfy0iL4jInvd+ND3tjoicnx6fF5Hb0++OnHP7IrI4/fyr+Gl+58Fwzn1SRD4pIuLyhfcVF5YPPp/gnpweZ0Z6PMnrcbavlfTjAr4QDqjczBDXOIYqdizWxzHvj7/vE7eRGaevkfqOw3d86u3jnjh//HyC75rnwXV5LfOcR/iLvN4E5/FZzRzz2om5N7eXIErgb4bzyvs41vPl4ufwuTOD+D2Ni/g8MfdmTQfPbNZ1LvE57oP3lOvqRUZlfVBeL3lt3l/qXfN5gr+l9oRLvffE/uBzpvaiuW7q94N7MmuIz8F9llpbief0qbXEwedM3YMcseaOIY84Z53tO1ve+zNv9Z1Hpjicc/Ny4C1cFZE9Efm/ROQjkVMP5yMm6vwRn9sPvP+ciHxORKSyfNE/+Xf/oYhY4Z9vUynod8clPafQ0HOGVXvp0q6+re6ivhEujGwPx0P9rWxfPx/U9HezA2zWin7O3yxv6nVb5/W6hX397uQIJVdd05XUOq8/bITXJP5bGTxDBfexe11/J9fR36HQHZX0+OK/2n5wfOdHF839UfFScVfX9cea53
[…]
