+++ v0.2.0/_sources/frequencies/dominant_frequencies.rst.txt
@@ -1,8 +1,0 @@
-spafe.frequencies.dominant_frequencies
-=======================================
-
-
-.. automodule:: spafe.frequencies.dominant_frequencies
-    :members:
-    :undoc-members:
-    :show-inheritance:
