+++ v0.1.2/_sources/frequencies/fundamental_frequencies.rst.txt
@@ -1,8 +1,0 @@
-spafe.frequencies.fundamental_frequencies
-==========================================
-
-
-.. automodule:: spafe.frequencies.fundamental_frequencies
-  :members:
-  :undoc-members:
-  :show-inheritance:
