+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,34 +1,0 @@
-bug report example
-==================
-
-#### Describe the bug
-
-A clear and concise description of what the bug is.
-
-#### To Reproduce
-
-Steps to reproduce the behavior:
-
-#### Code
-
-Include a minimal reproducible example, if possible.
-
-#### Expected behavior
-
-A clear and concise description of what you expected to happen.
-
-#### Screenshots
-
-If applicable, add screenshots to help explain your problem.
-
-#### System specifications
-
--	OS: (e.g. Ubunutu 18.04.3 LTS)
--	Python (e.g. python 3.5)
--	numpy (e.g. 1.17.2)
--	scipy (e.g. 1.3.1)
--	Spafe (e.g. 0.1.1)
-
-#### Additional context
-
-Add any other context about the problem here.

[…]
