### 2022-10-24 975efec
add fixes and edits recommended by hadware at #46
