### 2022-10-25 1941686
edit windowing and fixes suggested by hadware
