+++ spafe/features/gfcc.py
@@ -21,7 +21,7 @@ from ..utils.preprocessing import (
     framing,
     windowing,
     zero_handling,
-    WindowType,
+    SlidingWindow,
 )
 
 
@@ -30,9 +30,7 @@ def erb_spectrogram(
     fs: int = 16000,
     pre_emph: bool = True,
     pre_emph_coeff: float = 0.97,
-    win_len=0.025,
-    win_hop=0.01,
-    win_type: WindowType = "hamming",
+    window : Optional[SlidingWindow] = None,
     nfilts: int = 24,
     nfft: int = 512,
     low_freq: float = 0,
@@ -52,12 +50,8 @@ def erb_spectrogram(
                                       (Default is True).
         pre_emph_coeff      (float) : pre-emphasis filter coefficient.
                                       (Default is 0.97).
-        win_len             (float) : window length in sec.
-                                      (Default is 0.025).
-        win_hop             (float) : step between successive windows in sec.
-                                      (Default is 0.01).
-        win_type              (str) : window type to apply for the windowing.
-                                      (Default is "hamming".
+        window      (SlidingWindow) : sliding window object.
+                                      (Default is None).
         nfilts                (int) : the number of filters in the filter bank.
                                       (Default is 40).
         nfft                  (int) : number of FFT points.
@@ -93,10 +87,11 @@ def erb_spectrogram(
 
             from spafe.features.gfcc import erb_spectrogram
             from spafe.utils.vis import show_spectrogram
+            from spafe.utils.preprocessing import SlidingWindow
             from scipy.io.wavfile import read
 
             # read audio
-            fpath = "../../../test.wav"
+            fpath = "../../../data/test.wav"
             fs, sig = read(fpath)
 
             # compute erb spectrogram
@@ -104,9 +99,7 @@ def erb_spectrogram(
                                             fs=fs,
                                             pre_emph=0,
                                             pre_emph_coeff=0.97,
-                                            win_len=0.030,
-                                            win_hop=0.015,
-                                            win_type: WindowType="hamming",
+                                            window=SlidingWindow(0.03, 0.015, "hamming"),
                                             nfilts=128,
                                             nfft=2048,
                                             low_freq=0,
@@ -143,11 +136,15 @@ def erb_spectrogram(
     if pre_emph:
         sig = pre_emphasis(sig=sig, pre_emph_coeff=0.97)
 
+    # init window
+    if window is None:
+         window = SlidingWindow()
+
     # -> framing
-    frames, frame_length = framing(sig=sig, fs=fs, win_len=win_len, win_hop=win_hop)
+    frames, frame_length = framing(sig=sig, fs=fs, win_len=window.win_len, win_hop=window.win_hop)
 
     # -> windowing
-    windows = windowing(frames=frames, frame_len=frame_length, win_type=win_type)
+    windows = windowing(frames=frames, frame_len=frame_length, win_type=window.win_type)
 
     # -> FFT -> |.|
     ## Magnitude of the FFT
@@ -168,9 +165,7 @@ def gfcc(
     num_ceps: int = 13,
     pre_emph: bool = True,
     pre_emph_coeff: float = 0.97,
-    win_len: float = 0.025,
-    win_hop: float = 0.01,
-    win_type: WindowType = "hamming",
+    window : Optional[SlidingWindow] = None,
     nfilts: int = 24,
     nfft: int = 512,
     low_freq: float = 0,
@@ -197,12 +192,8 @@ def gfcc(
                                       (Default is True).
         pre_emph_coeff      (float) : pre-emphasis filter coefficient.
                                       (Default is 0.97).
-        win_len             (float) : window length in sec.
-                                      (Default is 0.025).
-        win_hop             (float) : step between successive windows in sec.
[…]
