+++ spafe/features/mfcc.py
@@ -21,21 +21,16 @@ from ..utils.preprocessing import (
     framing,
     windowing,
     zero_handling,
-    WindowType,
+    SlidingWindow,
 )
 
 
-# TODO: the mel conversion approaches are only ["Oshaghnessy", "Lindsay"]
-#  however the conversion functions has 3 : ["Oshaghnessy", "Beranek", "Lindsay"]
-#  is this on purpose?
 def mel_spectrogram(
     sig,
     fs: int = 16000,
     pre_emph: bool = True,
     pre_emph_coeff: float = 0.97,
-    win_len: float = 0.025,
-    win_hop: float = 0.01,
-    win_type: WindowType = "hamming",
+    window : Optional[SlidingWindow] = None,
     nfilts: int = 24,
     nfft: int = 512,
     low_freq: float = 0,
@@ -48,19 +43,15 @@ def mel_spectrogram(
     Compute the mel scale spectrogram.
 
     Args:
-        sig        (numpy.ndarray) : a mono audio signal (Nx1) from which to compute features.
-        fs                   (int) : the sampling frequency of the signal we are working with.
-                                     (Default is 16000).
-        pre_emph            (bool) : apply pre-emphasis if 1.
+        sig       (numpy.ndarray) : a mono audio signal (Nx1) from which to compute features.
+        fs                  (int) : the sampling frequency of the signal we are working with.
+                                    (Default is 16000).
+        pre_emph           (bool) : apply pre-emphasis if 1.
                                     (Default is True).
         pre_emph_coeff    (float) : pre-emphasis filter coefficient.
                                     (Default is 0.97).
-        win_len           (float) : window length in sec.
-                                    (Default is 0.025).
-        win_hop           (float) : step between successive windows in sec.
-                                    (Default is 0.01).
-        win_type            (str) : window type to apply for the windowing.
-                                    (Default is "hamming").
+        window    (SlidingWindow) : sliding window object.
+                                    (Default is None).
         nfilts              (int) : the number of filters in the filter bank.
                                     (Default is 40).
         nfft                (int) : number of FFT points.
@@ -84,7 +75,7 @@ def mel_spectrogram(
         - :code:`scale` : can take the following options ["constant", "ascendant", "descendant"].
         - :code:`dct` : can take the following options [1, 2, 3, 4].
         - :code:`normalize` : can take the following options ["mvn", "ms", "vn", "mn"].
-        - :code:`conversion_approach` : can take the following options ["Oshaghnessy", "Lindsay"].
+        - :code:`conversion_approach` : can take the following options ["Oshaghnessy", "Beranek", "Lindsay"].
           Note that the use of different options than the default can lead to unexpected behavior/issues.
 
     Note:
@@ -97,19 +88,18 @@ def mel_spectrogram(
 
             from spafe.features.mfcc import mel_spectrogram
             from spafe.utils.vis import show_spectrogram
+            from spafe.utils.preprocessing import SlidingWindow
             from scipy.io.wavfile import read
 
             # read audio
-            fpath = "../../../test.wav"
+            fpath = "../../../data/test.wav"
             fs, sig = read(fpath)
 
             mSpec, _ = mel_spectrogram(sig,
                                             fs=fs,
                                             pre_emph=0,
                                             pre_emph_coeff=0.97,
-                                            win_len=0.030,
-                                            win_hop=0.015,
-                                            win_type: WindowType="hamming",
+                                            window=SlidingWindow(0.03, 0.015, "hamming"),
                                             nfilts=128,
                                             nfft=2048,
[…]
