### 2022-11-05 #4739 @SuperKogito
Hello @faroit,
I implemented the changes requested in most of the issues and currently waiting for @hadware, who is aware of the changes and mentioned that he will review them asap. There is only one last issue ( no comparison with librosa); that I will deal with this weekend.
Thank you for your patience and efforts.
