### 2022-12-01 #4739 @faroit
@hadware @SuperKogito can you update us on the status of https://github.com/SuperKogito/spafe/pull/46?
