+++ features/pncc-1.py
@@ -1,9 +1,10 @@
 from scipy.io.wavfile import read
 from spafe.features.pncc import pncc
+from spafe.utils.preprocessing import SlidingWindow
 from spafe.utils.vis import show_features
 
 # read audio
-fpath = "../../../test.wav"
+fpath = "../../../data/test.wav"
 fs, sig = read(fpath)
 
 # compute pnccs
@@ -11,9 +12,7 @@ pnccs  = pncc(sig,
               fs=fs,
               pre_emph=0,
               pre_emph_coeff=0.97,
-              win_len=0.030,
-              win_hop=0.015,
-              win_type="hamming",
+              window=SlidingWindow(0.03, 0.015, "hamming"),
               nfilts=128,
               nfft=1024,
               low_freq=0,
