### 2022-12-30 84eab23
clean up docs and fix formatting
