+++ spafe/features/bfcc.py
@@ -51,7 +51,7 @@ def bark_spectrogram(
     fs: int = 16000,
     pre_emph: float = 0,
     pre_emph_coeff: float = 0.97,
-    window : Optional[SlidingWindow] = None,
+    window: Optional[SlidingWindow] = None,
     nfilts: int = 24,
     nfft: int = 512,
     low_freq: float = 0,
@@ -160,10 +160,12 @@ def bark_spectrogram(
 
     # init window
     if window is None:
-         window = SlidingWindow()
+        window = SlidingWindow()
 
     # -> framing
-    frames, frame_length = framing(sig=sig, fs=fs, win_len=window.win_len, win_hop=window.win_hop)
+    frames, frame_length = framing(
+        sig=sig, fs=fs, win_len=window.win_len, win_hop=window.win_hop
+    )
 
     # -> windowing
     windows = windowing(frames=frames, frame_len=frame_length, win_type=window.win_type)
@@ -186,7 +188,7 @@ def bfcc(
     num_ceps: int = 13,
     pre_emph: bool = True,
     pre_emph_coeff: float = 0.97,
-    window : Optional[SlidingWindow] = None,
+    window: Optional[SlidingWindow] = None,
     nfilts: int = 26,
     nfft: int = 512,
     low_freq: float = 0,
@@ -291,7 +293,6 @@ def bfcc(
     # run checks
     if nfilts < num_ceps:
         raise ParameterError(ErrorMsgs["nfilts"])
-
 
     # compute features
     features, fourrier_transform = bark_spectrogram(
