### 2022-12-30 #4739 @SuperKogito
First, let me thank you all @faroit  @hbredin  @hadware  for your efforts , assistance, guidance and mostly patience :heart:  I truly appreciate it.
I just merged all the changes and cleaned up a bit more my docs and Readme. I think all requested changes are now included in the lib.
I will be releasing the new version 0.2.1 at the latest next Monday, and I will adjust the paper to refer to that version.
@faroit  I think with that, my to-dos should be over right?
