### 2023-01-02 afbe42e
fix docs & version and update accordingly
