+++ spafe/features/cqcc.py
@@ -86,7 +86,7 @@ def cqt_spectrogram(
             from scipy.io.wavfile import read
 
             # read audio
-            fpath = "../../../data/test.wav"
+            fpath = "../../../tests/data/test.wav"
             fs, sig = read(fpath)
 
             # compute spectrogram
@@ -243,7 +243,7 @@ def cqcc(
             from spafe.utils.vis import show_features
 
             # read audio
-            fpath = "../../../data/test.wav"
+            fpath = "../../../tests/data/test.wav"
             fs, sig = read(fpath)
 
             # compute cqccs
