+++ spafe/features/msrcc.py
@@ -104,7 +104,7 @@ def msrcc(
             from spafe.utils.vis import show_features
 
             # read audio
-            fpath = "../../../data/test.wav"
+            fpath = "../../../tests/data/test.wav"
             fs, sig = read(fpath)
 
             # compute msrccs
