+++ spafe/features/ngcc.py
@@ -107,7 +107,7 @@ def ngcc(
             from spafe.utils.vis import show_features
 
             # read audio
-            fpath = "../../../data/test.wav"
+            fpath = "../../../tests/data/test.wav"
             fs, sig = read(fpath)
 
             # compute ngccs
