+++ docs/source/conf.py
@@ -25,9 +25,9 @@ author = "Ayoub Malek"
 html_favicon = "_static/favicon_io/favicon.ico"
 
 # The short X.Y version
-version = "0.2.0"
+version = "0.3.0"
 # The full version, including alpha/beta/rc tags
-release = "0.2.0"
+release = "0.3.0"
 
 # -- General configuration ---------------------------------------------------
 
