+++ spafe/frequencies/fundamental_frequencies.py
@@ -177,7 +177,7 @@ def compute_yin(
 
 
             # read audio
-            fpath = "../../../data/test.wav"
+            fpath = "../../../tests/data/test.wav"
             fs, sig = read(fpath)
             duration = len(sig) / fs
             harmonic_threshold = 0.85
