+++ spafe/features/psrcc.py
@@ -110,7 +110,7 @@ def psrcc(
             from spafe.utils.vis import show_features
 
             # read audio
-            fpath = "../../../data/test.wav"
+            fpath = "../../../tests/data/test.wav"
             fs, sig = read(fpath)
 
             # compute psrccs
