+++ features/lpc-2.py
@@ -4,7 +4,7 @@ from spafe.utils.preprocessing import SlidingWindow
 from spafe.utils.vis import show_features
 
 # read audio
-fpath = "../../../data/test.wav"
+fpath = "../../../tests/data/test.wav"
 fs, sig = read(fpath)
 
 # compute lpccs
