+++ frequencies/dominant_frequencies-1.py
@@ -9,7 +9,7 @@ win_len = 0.020
 win_hop = 0.010
 
 # read audio
-fpath = "../../../data/test.wav"
+fpath = "../../../tests/data/test.wav"
 fs, sig = read(fpath)
 
 # compute dominant frequencies
@@ -21,7 +21,7 @@ dominant_frequencies = get_dominant_frequencies(sig,
                                                 nfft=nfft,
                                                 win_len=win_len,
                                                 win_hop=win_hop,
-                                                win_type: WindowType="hamming")
+                                                win_type="hamming")
 
 # compute FFT, Magnitude, Power spectra
 fourrier_transform = np.absolute(np.fft.fft(sig, nfft))
