### 2023-01-03 #4739 @SuperKogito
Spafe  0.3.0 is now available on [Pypi](https://pypi.org/project/spafe/) and [anaconda](https://anaconda.org/SuperKogito/spafe).
The paper's latest version is [here](https://github.com/SuperKogito/spafe/blob/master/paper/paper.pdf),

Thank you all for this great learning experience, and happy new year !
