### 2023-01-10 #4739 @faroit
The submission looks good now and we can finally move forward 👏 I just did check the paper again for grammar and spellings and created a pull request that you can merge.

@SuperKogito, after the merge, when the paper is ready, please then please make another new release of the main repo and report the version number here.

Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the archive version, please make sure that:

- The authors of the archive are the same as the JOSS paper authors
