### 2023-01-12 196a0e9
update copyright and versions for paper
