+++ v0.3.0/features/spfeats-1.py
@@ -1,0 +1,11 @@
+from spafe.features.spfeats import extract_feats
+from scipy.io.wavfile import read
+from pprint import pprint
+
+# read audio
+fpath = "../../../data/test.wav"
+fs, sig = read(fpath)
+
+# compute erb spectrogram
+spectral_feats = extract_feats(sig, fs)
+pprint(spectral_feats)
