### 2023-01-13 #4739 @SuperKogito
- The latest release is  [0.3.1](https://github.com/SuperKogito/spafe/releases/tag/v0.3.1)
- Zenodo DOI: `10.5281/zenodo.10705`
  - [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10705.svg)](https://doi.org/10.5281/zenodo.10705)
