### 2023-01-13 #4739 @SuperKogito
- The latest release is [0.3.1](https://github.com/SuperKogito/spafe/releases/tag/v0.3.1)
- The Zenodo DOI is `10.5281/zenodo.7533946`
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7533946.svg)](https://doi.org/10.5281/zenodo.7533946)
