### 2023-01-13 #4739 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
