### 2023-01-13 #4739 @SuperKogito
isn't the DOI missing the first digit? it is supposed to be `10.5281/zenodo.7533946` and *not* `0.5281/zenodo.7533946`.
