### 2023-01-27 #4739 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
